Rename `util.py` to `core.py`
