+++ grafana_pandas_datasource/core.py
